feat(confirm) adicionado confirm global

+++ src/pages/usuarios/tabela/Actions.vue
@@ -4,14 +4,13 @@ import { Button } from '@/components/ui/button'
 import { DropdownMenu, DropdownMenuContent, DropdownMenuItem, DropdownMenuSeparator, DropdownMenuTrigger } from '@/components/ui/dropdown-menu'
 import type { Usuarios } from '@/types/schemas';
 import { useToast } from 'vue-toastification';
-import { AlertDialog, AlertDialogAction, AlertDialogCancel, AlertDialogContent, AlertDialogDescription, AlertDialogFooter, AlertDialogHeader, AlertDialogTitle } from '@/components/ui/alert-dialog';
 import { ref } from 'vue';
 import { useClientesStore } from '@/stores/clientes/useClientes';
 import { ClienteRepository } from '@/repositories/cliente-repository';
+import { useConfirm } from '@/composables/useConfirm';
 
 const store = useClientesStore()
 const openDelete = ref(false)
-const id = ref<number | null>(null)
 
 const { data } = defineProps<{
     data: Usuarios,
@@ -19,13 +18,14 @@ const { data } = defineProps<{
 
 const toast = useToast()
 
-function openDeleteModal(number: number) {
-    id.value = number
-    openDelete.value = true
-}
-
 async function deletar(id: number) {
     if (!id) return toast.error('ID não informado!')
+    const confirm = await useConfirm().confirm({
+        title: 'Excluir cliente',
+        message: 'Tem certeza que deseja excluir este cliente?',
+        confirmText: 'Sim, excluir!',
+    })
+    if (!confirm) return
     try {
         await ClienteRepository.remove(id)
         store.updateTable()
@@ -40,24 +40,6 @@ async function deletar(id: number) {
 </script>
 
 <template>
-
-    <AlertDialog :open="openDelete">
-        <AlertDialogContent>
-            <AlertDialogHeader>
-                <AlertDialogTitle>Deseja deletar o registro?</AlertDialogTitle>
-                <AlertDialogDescription>
-                    Essa ação não poderá ser desfeita, o registro sera deletado permanentemente
-                </AlertDialogDescription>
-            </AlertDialogHeader>
-            <AlertDialogFooter>
-                <AlertDialogCancel @click="openDelete = false">Cancelar</AlertDialogCancel>
-                <AlertDialogAction class="text-white bg-danger hover:bg-danger/40" @click="deletar(id!)">
-                    Sim, deletar
-                </AlertDialogAction>
-            </AlertDialogFooter>
-        </AlertDialogContent>
-    </AlertDialog>
-
     <DropdownMenu>
         <DropdownMenuTrigger as-child>
             <Button variant="default" class="w-8 h-8 p-0 text-white">
@@ -71,7 +53,7 @@ async function deletar(id: number) {
                 Editar
             </DropdownMenuItem>
             <DropdownMenuSeparator />
-            <DropdownMenuItem class="text-danger" @click="openDeleteModal(data.id!)">
+            <DropdownMenuItem class="text-danger" @click="deletar(data.id!)">
                 <i class="fa-regular fa-trash-can mr-1"></i>
                 Excluir
             </DropdownMenuItem>
